+++ packages/provider/src/index.ts
@@ -1,4 +1,4 @@
-// Copyright 2021-2025 Prosopo (UK) Ltd.
+// Copyright 2021-2026 Prosopo (UK) Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -29,3 +29,4 @@ export * from "./api/ignoreMiddleware.js";
 export * from "./api/robotsMiddleware.js";
 export * from "./compositeIpAddress.js";
 export * from "./services/ipComparison.js";
+export * from "./api/blacklistRequestInspector.js";